v3 comp for remote debugging
