+++ PowerShellTools/DependencyValidator.cs
@@ -4,6 +4,7 @@ using EnvDTE;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.Win32;
 using System.ComponentModel.Composition;
+using PowerShellTools.Common;
 
 namespace PowerShellTools
 {
@@ -62,29 +63,7 @@ namespace PowerShellTools
         {
             get
             {
-                var version = new Version(0, 0);
-                using (var reg = Registry.LocalMachine.OpenSubKey(@"Software\Microsoft\PowerShell\3\PowerShellEngine"))
-                {
-                    if (reg != null)
-                    {
-                        var versionString = reg.GetValue("PowerShellVersion") as string;
-
-                        Version.TryParse(versionString, out version);
-                        return version;
-                    }
-                }
-
-                using (var reg = Registry.LocalMachine.OpenSubKey(@"Software\Microsoft\PowerShell\1\PowerShellEngine"))
-                {
-                    if (reg != null)
-                    {
-                        var versionString = reg.GetValue("PowerShellVersion") as string;
-                        Version.TryParse(versionString, out version);
-                        return version;
-                    }
-                }
-
-                return version;
+                return DependencyUtilities.GetInstalledPowerShellVersion();
             }
         }
     }
